cv_matcher/requirements.txt
# Extraction de texte PDF
pymupdf==1.23.1
pdfminer.six==20221105

# NLP et ML
transformers==4.31.0
sentence-transformers==2.2.2
spacy==3.6.1
scikit-learn==1.3.0
nltk==3.8.1

# API
fastapi==0.103.1
uvicorn==0.23.2
pydantic==2.3.0

# Utilitaires
python-multipart==0.0.6